Week 3 update
Changed position of title to be within the head tag and changed the multiple h1 tags to h2 tags so there is only one title.

<b> items.
Created <div class""> for sections and classed them. Italicized some sections using <i>.
Added in link to instagram in reference to myself with <a href>

Added clarification on the recipe to provide an easier and clear understanding. Also updated ingredients to include water and added in note for a simple side dish within the stock section.