Updated index and image
Updated index to have more information on meat to match the photo that has been added. Image was routed to the right folder to show up online. Style sheet updated to contain font weight and size for the typography section of the page. also updated the layout within CSS to adjust to how I wanted the image to be.